Add an additional option for closedPos.
closedPos: "absolute",// String: Position of the closed nav.

Works the same as openPos, prevents from needing to use !important in CSS or modify the JS library if you want to set an alternative default position.

Also, to get the positions entirely from CSS, both options can be set to "". I didn't make this the default to ensure backwards compatibility with existing sites.

I've also updated readme.md to include the new option.

I've been working with responsive-nav in a project and need to modify it to make it flexible enough to fit my needs. Hoping to push the changes back into the project for everyone else's benefit.

Signed-off-by: Jonathan Haslett <jon@dashmedia.com.au>