Scope css so it doesn't mess sites up
As soon as somebody drops this into their site it will affect all buttons. It might go unnoticed and cause problems. 

The button has an ID so this change will prevent the demo "leaking" when dropped into a real site. I normally stay away from ids but there should only be one and the current css is never going to be used on a live site either so you might modify it either way around but this one stops the leak.